==> 1_feature_engineering/02_Custom features.ipynb <==
## In this kernel I would like to show: 
## 1. FE creation approaches
## 2. Sequential fe validation
## 3. Dimension reduction
## 4. FE validation by Permutation importance
## 5. Mean encodings
## 6. Parallelization for FE

import numpy as np 
import pandas as pd 
import os, sys, gc, warnings, psutil, random

warnings.filterwarnings('ignore')

def timeseries_target_encoding(dt, year_cols, cols, func):
    year_target_dict = pd.DataFrame()
    for i, year in enumerate(year_cols):            
        if i  >= len(year_cols) - 1:
            break
        dt_group_tmp = dt[dt["year"] <= year].groupby(cols).agg(
            timeseries_target_encoding = ("sales", func)
        ).reset_index()
        dt_group_tmp = dt_group_tmp.rename(columns={"timeseries_target_encoding": f'enc_timeseries_{"_".join(cols)}_{func}'})
        year_target_dict = dt_group_tmp
        year_target_dict['year'] = year + 1
    dt = pd.merge(dt, year_target_dict, how="left", on=["year"] + cols)
    return dt
year_cols = [2011, 2012, 2013, 2014, 2015, 2016]

########################### Apply on grid_df
#################################################################################
# lets read grid from 
# https://www.kaggle.com/kyakovlev/m5-simple-fe
# to be sure that our grids are aligned by index

# データのパス変更の必要あり
# grid_df = pd.read_pickle('../data/preprocess/grid_part_1.pkl')
# calendar = pd.read_csv("../data/raw/calendar.csv")
TARGET = 'sales' 
grid_df = pd.read_pickle('../data/preprocess/grid_part_1_update.pkl')
calendar = pd.read_csv("../data/raw/calendar.csv")

calendar["d"] = calendar["d"].apply(lambda x : x.split("_")[1])
calendar["d"] = calendar["d"].astype("int16")
grid_df = pd.merge(grid_df, calendar[["d", "year"]], how="left", on="d")

grid_df[TARGET][grid_df['d']>(1913-28)] = np.nan
grid_df.loc[grid_df[TARGET]==0, TARGET] = np.nan
base_cols = list(grid_df)

icols =  [
            ['state_id'],
            ['store_id'],
            ['cat_id'],
            ['dept_id'],
            ['state_id', 'cat_id'],
            ['state_id', 'dept_id'],
            ['store_id', 'cat_id'],
            ['store_id', 'dept_id'],
            ['item_id'],
            ['item_id', 'state_id'],
            ['item_id', 'store_id']
            ]
"""
for col in icols:
    print('Encoding', col)
    col_name = '_'+'_'.join(col)+'_'
    grid_df['enc'+col_name+'mean'] = grid_df.groupby(col)[TARGET].transform('mean').astype(np.float16)
    grid_df['enc'+col_name+'std'] = grid_df.groupby(col)[TARGET].transform('std').astype(np.float16)
"""
for cols in icols:
    print('Encoding', cols)
    grid_df = timeseries_target_encoding(grid_df, year_cols, cols, "mean")
    grid_df = timeseries_target_encoding(grid_df, year_cols, cols, "std")



keep_cols = [col for col in list(grid_df) if col not in base_cols]
grid_df = grid_df[['id','d']+keep_cols]

keep_cols

for col in keep_cols:
    grid_df[col] = grid_df[col].astype("float32")

grid_df

#################################################################################
print('Save Mean/Std encoding')
grid_df.to_pickle('../data/preprocess/mean_encoding_df_update_nonleak.pkl')

########################### Final list of new features
#################################################################################
grid_df.info()
