==> 1_feature_engineering/03_Lag features.ipynb <==
# General imports
import numpy as np
import pandas as pd
import os, sys, gc, time, warnings, pickle, psutil, random

import time

warnings.filterwarnings('ignore')

## Simple "Memory profilers" to see memory usage
def get_memory_usage():
    return np.round(psutil.Process(os.getpid()).memory_info()[0]/2.**30, 2) 
        
def sizeof_fmt(num, suffix='B'):
    for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
        if abs(num) < 1024.0:
            return "%3.1f%s%s" % (num, unit, suffix)
        num /= 1024.0
    return "%.1f%s%s" % (num, 'Yi', suffix)

########################### Vars
#################################################################################
TARGET = 'sales'         # Our main target
END_TRAIN = 1913+28         # Last day in train set
MAIN_INDEX = ['id','d']  # We can identify item by these columns

%%time
for i in range(1, 1+1):
    
    grid_df = pd.read_pickle('./grid_part_1_update.pkl')

    # We need only 'id','d','sales'
    # to make lags and rollings
    grid_df = grid_df[['id','d','sales']]
    
    SHIFT_DAY = i
    
    start_time = time.time()
    print('Create lags ' + str(SHIFT_DAY))
    
    LAG_DAYS = [col for col in range(SHIFT_DAY, SHIFT_DAY+15)]
    grid_df = grid_df.assign(**{
            '{}_lag_{}'.format(col, l): grid_df.groupby(['id'])[col].transform(lambda x: x.shift(l))
            for l in LAG_DAYS
            for col in [TARGET]
        })
    
    # Minify lag columns
    for col in list(grid_df):
        if 'lag' in col:
            grid_df[col] = grid_df[col].astype(np.float16)
    
    print('%0.2f min: Lags' % ((time.time() - start_time) / 60))
    
    """start_time = time.time()
    print('Create rolling aggs')

    for i in [7,14,21,28,35,42,49,60,180]:
        print('Rolling period:', i)
        grid_df['rolling_mean_'+str(i)] = grid_df.groupby(['id'])[TARGET].transform(lambda x: x.shift(SHIFT_DAY).rolling(i).mean()).astype(np.float16)
        grid_df['rolling_std_'+str(i)]  = grid_df.groupby(['id'])[TARGET].transform(lambda x: x.shift(SHIFT_DAY).rolling(i).std()).astype(np.float16)
    """
    # Rollings
    # with sliding shift
    for d_shift in [SHIFT_DAY, SHIFT_DAY+2, SHIFT_DAY+6, SHIFT_DAY+13]: 
        print('Shifting period:', d_shift)
        for d_window in [7,14,28,60,180]:
            col_name = 'rolling_mean_tmp_'+str(d_shift)+'_'+str(d_window)
            grid_df[col_name] = grid_df.groupby(['id'])[TARGET].transform(lambda x: x.shift(d_shift).rolling(d_window).mean()).astype(np.float16)


    print('%0.2f min: Lags' % ((time.time() - start_time) / 60))
        
    ########################### Export
    #################################################################################
    print('Save lags and rollings')
    grid_df.to_pickle('lags_df_'+str(SHIFT_DAY)+'_dbd.pkl')

grid_df.columns
