==> 1_feature_engineering/05_Week weighted moving average.ipynb <==
# General imports
import numpy as np
import pandas as pd
import os, sys, gc, time, warnings, pickle, psutil, random

import time

warnings.filterwarnings('ignore')

########################### Vars
#################################################################################
TARGET = 'sales'         # Our main target
END_TRAIN = 1913+28         # Last day in train set
MAIN_INDEX = ['id','d']  # We can identify item by these columns

grid_df = pd.read_pickle('./grid_part_1_update.pkl')
plus_df = pd.read_pickle('./grid_part_3_update.pkl')
grid_df = pd.concat([grid_df, plus_df.drop(["d", "id"], axis=1)], axis=1)
tmp = grid_df[['id','d', 'store_id', 'sales', 'tm_y', 'tm_m', 'tm_d', 'tm_dw', "tm_w"]].copy()

tmp.head()

tmp["tm_w"].describe()

tmp["tm_dw"].describe()

## tm_dw=0を集める

tmp2 = tmp[tmp["tm_dw"]==0]

tmp2[tmp2["id"]=="HOBBIES_1_008_CA_1_evaluation"].head()

# 24週前までの月曜日のsalesを横付けしていく
for i in range(1, 25):
    print(i)
    tmp2["lag_"+str(i)+"_week"] = tmp2.groupby(['id'])["sales"].transform(lambda x: x.shift(i)).astype(np.float16)

tmp2[tmp2["id"]=="HOBBIES_1_008_CA_1_evaluation"].head(6)

tmp2[tmp2["id"]=="HOBBIES_1_008_CA_1_evaluation"].tail(6)

- これで月曜日(tm_dw=0)に関しては、lag_n_week列を集約すれば移動平均が求められるようになった。
- 注意点として、28日前までの情報は使わないことにしているので、実際はlag_5_week移行の列で平均などを取らないといけない
- 元のデータフレームとは tm_y, tm_w の複合キーが一意にある週を示すので、これとidでmergeすればよい
- ※tm_y：年、tm_w：一年の何週目か(1～53で、多分4/1週が1,3/31週が53)

# 5週前から4週間までさかのぼった時の移動平均
values_5week_4week = [tmp2["lag_"+str(5+i)+"_week"].values for i in range(4)]
tmp2["ave_5week_4week_tm_dw0"] = np.average(values_5week_4week,  axis=0)

# 5週前から8週間までさかのぼった時の移動平均
values_5week_8week = [tmp2["lag_"+str(5+i)+"_week"].values for i in range(8)]
tmp2["ave_5week_8week_tm_dw0"] = np.average(values_5week_8week,  axis=0)

# 5週前から12週間までさかのぼった時の移動平均
values_5week_12week = [tmp2["lag_"+str(5+i)+"_week"].values for i in range(12)]
tmp2["ave_5week_12week_tm_dw0"] = np.average(values_5week_12week,  axis=0)

tmp2

tmp2.to_pickle("tm_dw_0_df_update.pkl")

## 1~6についても移動平均作る

%%time

for j in range(1,7):
    print(j)
    
    tmp2 = tmp[tmp["tm_dw"]==j]
    
    # 24週前までの月曜日のsalesを横付けしていく
    for i in range(1, 25):
        # print(i)
        tmp2["lag_"+str(i)+"_week"] = tmp2.groupby(['id'])["sales"].transform(lambda x: x.shift(i)).astype(np.float16)
        
    # 5週前から4週間までさかのぼった時の移動平均
    values_5week_4week = [tmp2["lag_"+str(5+i)+"_week"].values for i in range(4)]
    tmp2["ave_5week_4week_tm_dw"+str(j)] = np.average(values_5week_4week,  axis=0)
    
    # 5週前から8週間までさかのぼった時の移動平均
    values_5week_8week = [tmp2["lag_"+str(5+i)+"_week"].values for i in range(8)]
    tmp2["ave_5week_8week_tm_dw"+str(j)] = np.average(values_5week_8week,  axis=0)
    
    # 5週前から12週間までさかのぼった時の移動平均
    values_5week_12week = [tmp2["lag_"+str(5+i)+"_week"].values for i in range(12)]
    tmp2["ave_5week_12week_tm_dw"+str(j)] = np.average(values_5week_12week,  axis=0)
    
    tmp2.to_pickle("tm_dw_" + str(j) +"_df.pkl")

df = pd.read_pickle("tm_dw_0_df_update.pkl")

df.tail()

df = df[["id", "tm_y", "tm_w", 'ave_5week_4week_tm_dw0', 'ave_5week_8week_tm_dw0', 'ave_5week_12week_tm_dw0']]

df.tail()

for i in range(1, 7):
    tmp = pd.read_pickle("tm_dw_{}_df.pkl".format(i))
    feat = ["id", "tm_y", "tm_w", 
            'ave_5week_4week_tm_dw'+str(i), 
            'ave_5week_8week_tm_dw'+str(i),
            'ave_5week_12week_tm_dw'+str(i)]
    tmp = tmp[feat]
    df = df.merge(tmp, on=["id", "tm_y", "tm_w"], how="left")
    

df

df.to_pickle("week_moving_average.pkl")

tmp.tail()

## 加重平均

weight = [(0.9)**i for i in range(4)]

weight

df = pd.read_pickle("tm_dw_0_df_update.pkl")

# 5週前から4週間までさかのぼった時の移動平均
weight = [(0.9)**i for i in range(4)]
values_5week_4week = [df["lag_"+str(5+i)+"_week"].values for i in range(4)]
df["ave_5week_4week_weight_tm_dw0"] = np.average(values_5week_4week,  weights=weight, axis=0)

# 5週前から8週間までさかのぼった時の加重平均
weight = [(0.9)**i for i in range(8)]
values_5week_8week = [df["lag_"+str(5+i)+"_week"].values for i in range(8)]
df["ave_5week_8week_weight_tm_dw0"] = np.average(values_5week_8week,  weights=weight, axis=0)

# 5週前から12週間までさかのぼった時の加重平均
weight = [(0.9)**i for i in range(12)]
values_5week_12week = [df["lag_"+str(5+i)+"_week"].values for i in range(12)]
df["ave_5week_12week_weight_tm_dw0"] = np.average(values_5week_12week,  weights=weight, axis=0)

df[df["id"]=="HOBBIES_1_008_CA_1_evaluation"].tail(6)

df.to_pickle("tm_dw_0_df.pkl")

for k in range(1, 7):
    df = pd.read_pickle("tm_dw_{}_df.pkl".format(k))
    
    # 5週前から4週間までさかのぼった時の移動平均
    weight = [(0.9)**i for i in range(4)]
    values_5week_4week = [df["lag_"+str(5+i)+"_week"].values for i in range(4)]
    df["ave_5week_4week_weight_tm_dw"+str(k)] = np.average(values_5week_4week,  weights=weight, axis=0)
    
    # 5週前から8週間までさかのぼった時の加重平均
    weight = [(0.9)**i for i in range(8)]
    values_5week_8week = [df["lag_"+str(5+i)+"_week"].values for i in range(8)]
    df["ave_5week_8week_weight_tm_dw"+str(k)] = np.average(values_5week_8week,  weights=weight, axis=0)
    
    # 5週前から12週間までさかのぼった時の加重平均
    weight = [(0.9)**i for i in range(12)]
    values_5week_12week = [df["lag_"+str(5+i)+"_week"].values for i in range(12)]
    df["ave_5week_12week_weight_tm_dw"+str(k)] = np.average(values_5week_12week,  weights=weight, axis=0)
    
    df.to_pickle("tm_dw_{}_df.pkl".format(k))

df = pd.read_pickle("tm_dw_0_df.pkl")

df = df[["id", "tm_y", "tm_w", 'ave_5week_4week_weight_tm_dw0',
         'ave_5week_8week_weight_tm_dw0', 'ave_5week_12week_weight_tm_dw0']]

for i in range(1, 7):
    tmp = pd.read_pickle("tm_dw_{}_df.pkl".format(i))
    feat = ["id", "tm_y", "tm_w", 
            'ave_5week_4week_weight_tm_dw'+str(i), 
            'ave_5week_8week_weight_tm_dw'+str(i),
            'ave_5week_12week_weight_tm_dw'+str(i)]
    tmp = tmp[feat]
    df = df.merge(tmp, on=["id", "tm_y", "tm_w"], how="left")

df.head()

df.to_pickle("week_w_moving_average.pkl")
