==> 1_feature_engineering/09_Next day is holiday.ipynb <==
holiday, workingdayに関連した特徴量をpickleファイルとしてdumpするnotebook  
メモリを抑える処理などは次のnotebookから拝借した https://www.kaggle.com/kyakovlev/m5-simple-fe

dumpされたpickleファイルのkeyは[id, d]

import numpy as np
import pandas as pd
import datetime
from datetime import date
import os, sys, gc, time, warnings, pickle, psutil, random

!pip install workalendar==8.4.0

from workalendar.usa import California
from workalendar.usa import Texas
from workalendar.usa import Wisconsin

pd.options.display.max_columns = 100
pd.options.display.max_rows = 100

## Simple "Memory profilers" to see memory usage
def get_memory_usage():
    return np.round(psutil.Process(os.getpid()).memory_info()[0]/2.**30, 2) 
        
def sizeof_fmt(num, suffix='B'):
    for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
        if abs(num) < 1024.0:
            return "%3.1f%s%s" % (num, unit, suffix)
        num /= 1024.0
    return "%.1f%s%s" % (num, 'Yi', suffix)

## Memory Reducer
# :df pandas dataframe to reduce size             # type: pd.DataFrame()
# :verbose                                        # type: bool
def reduce_mem_usage(df, verbose=True):
    numerics = ['int16', 'int32', 'int64', 'float16', 'float32', 'float64']
    start_mem = df.memory_usage().sum() / 1024**2    
    for col in df.columns:
        col_type = df[col].dtypes
        if col_type in numerics:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                       df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)    
    end_mem = df.memory_usage().sum() / 1024**2
    if verbose: print('Mem. usage decreased to {:5.2f} Mb ({:.1f}% reduction)'.format(end_mem, 100 * (start_mem - end_mem) / start_mem))
    return df

def make_holiday(df):
    
    def get_holiday(year, month, day, state_id):
        if state_id == "CA":
            return 1 if California().is_holiday(date(year, month, day)) else 0
        elif state_id == "TX":
            return 1 if Texas().is_holiday(date(year, month, day)) else 0
        elif state_id == "WI":
            return 1 if Wisconsin().is_holiday(date(year, month, day)) else 0
        else:
            return 0
        
    def get_working_day(year, month, day, state_id):
        if state_id == "CA":
            return 1 if California().is_working_day(date(year, month, day)) else 0
        elif state_id == "TX":
            return 1 if Texas().is_working_day(date(year, month, day)) else 0
        elif state_id == "WI":
            return 1 if Wisconsin().is_working_day(date(year, month, day)) else 0
        else:
            return 0
        
    tmp = df.drop_duplicates(subset=["date", "state_id"])
    
    years  = tmp["date"].dt.year.values
    months = tmp["date"].dt.month.values
    days   = tmp["date"].dt.day.values
    states = tmp["state_id"].values
    
    holidays = []
    workingdays = []
    
    for i in range(len(years)):
        holidays.append(get_holiday(years[i], months[i], days[i], states[i]))
        workingdays.append(get_working_day(years[i], months[i], days[i], states[i]))
        
    # df["is_holiday"] = df.apply(lambda x: get_holiday(x["year"], x["month"], x["mday"], x["state_id"]), axis=1)
    tmp["is_holiday"] = holidays
    tmp["is_workingday"] = workingdays
    for_merge = tmp[["date", "state_id", "is_holiday", "is_workingday"]]
    df = pd.merge(df, for_merge, on=["date", "state_id"], how="left")
    
    return df

def days_to_next_holiday(df):
    cal_holidays = {}
    for state_id in ["CA", "TX", "WI"]:
        if state_id == "CA":
            cal = California()
        elif state_id == "TX":
            cal = Texas()
        elif state_id == "WI":
            cal = Wisconsin()
        else:
            raise RuntimeError
        
        holidays = []
        for year in range(2011, 2017):
            for holiday in cal.holidays(year):
                if holiday[0] in holidays:
                    continue
                holidays.append(holiday[0])
        
        cal_holidays[state_id] = holidays
    
    tmp = df.drop_duplicates(subset=["date", "state_id"])
    days_to_next_holiday = []
    for i, row in tmp.iterrows():
        d = row["date"].date()
        holidays = cal_holidays[row["state_id"]]
        for h in range(len(holidays) - 1):
            if holidays[h + 1] == d:
                days_to_next_holiday.append(int((holidays[h + 2] - d).days))
                break
            elif holidays[h] <= d < holidays[h + 1]:
                days_to_next_holiday.append(int((holidays[h + 1] - d).days))
                break
            else:
                pass

    assert len(tmp) == len(days_to_next_holiday), print(len(tmp), len(days_to_next_holiday))
    tmp["days_to_next_holiday"] = days_to_next_holiday
    for_merge = tmp[["date", "state_id", "days_to_next_holiday"]]
    df = pd.merge(df, for_merge, on=["date", "state_id"], how="left")
    
    return df

def days_from_prev_holiday(df):
    cal_holidays = {}
    for state_id in ["CA", "TX", "WI"]:
        if state_id == "CA":
            cal = California()
        elif state_id == "TX":
            cal = Texas()
        elif state_id == "WI":
            cal = Wisconsin()
        else:
            raise RuntimeError
        
        holidays = []
        for year in range(2011, 2017):
            for holiday in cal.holidays(year):
                if holiday[0] in holidays:
                    continue
                holidays.append(holiday[0])
        
        cal_holidays[state_id] = holidays
    
    tmp = df.drop_duplicates(subset=["date", "state_id"])
    days_from_prev_holiday = []
    for i, row in tmp.iterrows():
        d = row["date"].date()
        holidays = cal_holidays[row["state_id"]]
        for h in range(1, len(holidays)):
            if holidays[h - 1] == d:
                days_from_prev_holiday.append(int((d - holidays[h - 2]).days))
                break
            elif holidays[h - 1] < d <= holidays[h]:
                days_from_prev_holiday.append(int((d - holidays[h - 1]).days))
                break
            else:
                pass

    assert len(tmp) == len(days_from_prev_holiday), print(len(tmp), len(days_from_prev_holiday))
    tmp["days_from_prev_holiday"] = days_from_prev_holiday
    for_merge = tmp[["date", "state_id", "days_from_prev_holiday"]]
    df = pd.merge(df, for_merge, on=["date", "state_id"], how="left")
    
    return df

def near_day_is_holiday(df):
    def check_next_day_is_holiday(row):
        state_id = row['state_id']
        if state_id == "CA":
            cal = California()
        elif state_id == "TX":
            cal = Texas()
        elif state_id == "WI":
            cal = Wisconsin()
        else:
            raise RuntimeError
                
        d = row["date"]
        next_day = d + datetime.timedelta(days=1)
        return 1 if cal.is_holiday(next_day) else 0
    
    def check_prev_day_is_holiday(row):
        state_id = row['state_id']
        if state_id == "CA":
            cal = California()
        elif state_id == "TX":
            cal = Texas()
        elif state_id == "WI":
            cal = Wisconsin()
        else:
            raise RuntimeError
                
        d = row["date"]
        prev_day = d - datetime.timedelta(days=1)
        return 1 if cal.is_holiday(prev_day) else 0
    
    tmp = df.drop_duplicates(subset=["date", "state_id"])
    tmp['next_day_is_holiday'] = tmp.apply(check_next_day_is_holiday, axis=1)
    tmp['prev_day_is_holiday'] = tmp.apply(check_prev_day_is_holiday, axis=1)
    
    for_merge = tmp[["date", "state_id", "next_day_is_holiday", "prev_day_is_holiday"]]
    df = pd.merge(df, for_merge, on=["date", "state_id"], how="left")
    
    return df

def monday_or_friday(df):
    tmp = df.drop_duplicates(subset=["date", "wday"])
    tmp['monday_or_friday'] = tmp["wday"].apply(lambda x: 1 if x == 3 or x == 7 else 0)
    
    for_merge = tmp[["date", "monday_or_friday"]]
    df = pd.merge(df, for_merge, on=["date"], how="left")
    
    return df

train_df = pd.read_csv('./sales_train_evaluation.csv')
calendar_df = pd.read_csv("../data/calendar.csv")

train_df.head()

calendar_df.head()

TARGET = 'sales' 
index_columns = ['id', 'item_id', 'dept_id', 'cat_id', 'store_id', 'state_id']
grid_df = pd.melt(train_df, 
                  id_vars=index_columns, 
                  var_name='d', 
                  value_name=TARGET)
del train_df

# Let's check our memory usage
print("{:>20}: {:>8}".format('Original grid_df', sizeof_fmt(grid_df.memory_usage(index=True).sum())))

# We can free some memory 
# by converting "strings" to categorical
# it will not affect merging and 
# we will not lose any valuable data
for col in index_columns:
    grid_df[col] = grid_df[col].astype('category')

# Let's check again memory usage
print("{:>20}: {:>8}".format('Reduced grid_df',sizeof_fmt(grid_df.memory_usage(index=True).sum())))

aaa = pd.read_pickle("../20200602_追加データの確認/grid_part_1_update.pkl")

aaa

grid_df_2 = aaa[["id", "state_id", "d"]].copy()

grid_df_2

bbb = grid_df_2["d"].values

result = []

for b in bbb:
    result.append("d_"+str(b))

grid_df_2["d"] = result

grid_df_2

grid_df = grid_df[["id", "state_id", "d"]]

grid_df = pd.concat([grid_df, grid_df_2])

grid_df = grid_df.drop_duplicates(["id", "state_id", "d"])

grid_df.shape

30490 * 1969

grid_df = pd.merge(grid_df[["id", "state_id", "d"]], calendar_df[["date", "wday", "month", "year", "d"]], how="left", on="d")
grid_df['date'] = pd.to_datetime(grid_df['date'], format='%Y-%m-%d')
grid_df = reduce_mem_usage(grid_df)
del calendar_df

grid_df.tail()

grid_df = make_holiday(grid_df)

%%time

grid_df = days_to_next_holiday(grid_df)
grid_df.head()

%%time

grid_df = days_from_prev_holiday(grid_df)
grid_df.head()

California().holidays(2011)

print(grid_df['days_to_next_holiday'].unique())
print(sorted(grid_df['days_to_next_holiday'].value_counts().items()))

grid_df[grid_df["days_to_next_holiday"] == 133]

Wisconsin().holidays(2012)

%%time

grid_df = near_day_is_holiday(grid_df)
grid_df.head()

grid_df = monday_or_friday(grid_df)
grid_df.head()

%%time

feature_columns = ["id", "d",
                   "is_holiday",
                   "is_workingday",
                   "days_to_next_holiday",
                   "days_from_prev_holiday",
                   "next_day_is_holiday",
                   "prev_day_is_holiday",
                   "monday_or_friday"]
grid_df = grid_df[feature_columns]

grid_df['d'] = grid_df['d'].apply(lambda x: x[2:]).astype(np.int16)

grid_df["id"] = grid_df["id"].astype('category')
grid_df["is_holiday"] = grid_df["is_holiday"].astype('category')
grid_df["is_workingday"] = grid_df["is_workingday"].astype('category')
grid_df["next_day_is_holiday"] = grid_df["next_day_is_holiday"].astype('category')
grid_df["prev_day_is_holiday"] = grid_df["prev_day_is_holiday"].astype('category')
grid_df["monday_or_friday"] = grid_df["monday_or_friday"].astype('category')

grid_df["days_to_next_holiday"] = grid_df["days_to_next_holiday"].astype(np.int8) 
grid_df["days_from_prev_holiday"] = grid_df["days_from_prev_holiday"].astype(np.int8) 

grid_df.to_pickle('holiday_features.pkl')
